+++ lib/src/helper/navigator_observer.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_smart_dialog/flutter_smart_dialog.dart';
+import 'package:flutter_smart_dialog/src/smart_dialog.dart';
 
 import 'dialog_proxy.dart';
 
@@ -17,6 +18,6 @@ class SmartNavigatorObserver extends NavigatorObserver {
   void didPop(Route<dynamic> route, Route<dynamic>? previousRoute) async {
     if (!SmartDialog.config.isExist) return;
 
-    DialogProxy.instance.closeAllDialog(pop: true);
+    DialogProxy.instance.dismiss(status: SmartStatus.allDialog);
   }
 }